pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bsccm-i2i"
version = "0.1.0"
description = "Bootstrap package for BSCCM image-to-image experiments."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "hydra-core>=1.3.2",
  "pydantic>=2.8.0",
  "typer>=0.12.0",
  "python-dotenv>=1.0.1",
  "pytorch-lightning>=2.2.0",
  "tensorboard>=2.14.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.0.0",
  "mypy>=1.10.0",
  "pytest>=8.0.0",
  "ruff>=0.4.0",
]

[project.scripts]
bsccm-i2i = "bsccm_i2i.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
